requirements/base.txt
bpython==0.12
Django==1.6.2
django-braces==1.2.2
django-model-utils==1.5.0
django-simple-captcha==0.4.2
logutils==0.3.3
South==0.8.2
django-taggit==0.12.2
